// src/dashboard/app/styles/exports.scss.d.ts
export interface ScssVariables {
    mainGradientStart: string;
    mainGradientStartLight: string;
    mainGradientEnd: string;
    cardGradientStart: string;
    cardGradientEnd: string;
    sequentialPalette: string;
    qualitativePalette1: string;
    qualitativePalette2: string;
    qualitativePalette3: string;
    text: string;
    fontFamily: string;
}

export const scssVariables: ScssVariables;

export default scssVariables;